refactor: done part3
