+++ analysis_options.yaml
@@ -2,6 +2,7 @@ analyzer:
   errors:
     avoid_dynamic_calls: ignore
     avoid_positional_boolean_parameters: ignore
+    avoid_print: ignore
     cascade_invocations: ignore
     deprecated_member_use: ignore
     inference_failure_on_instance_creation: ignore
