+++ pubspec.yaml
@@ -38,29 +38,23 @@ dev_dependencies:
   mocktail: ^1.0.4
   very_good_analysis: ^7.0.0
 
-
-# flutter_icons:
-#   android: true
-#   ios: true
-#   image_path: "lib/assets/png/logo_app.png"
-
 flutter:
   uses-material-design: true
   generate: true
   assets:
-  - .env
-  - lib/assets/icon/
-  - lib/assets/svg/
-  - lib/assets/icon/active_navbar/
-  - lib/assets/icon/inactive_navbar/
-  - lib/assets/icon/home_icon/
-  - lib/assets/icon/add_icon/
-  - lib/assets/icon/components_icon/
-  - lib/assets/icon/popup_icon/
-  - lib/assets/icon/setting_icon/
-  - lib/assets/icon/transaction_icon/
-  - lib/assets/icon/figma_svg/
-  - lib/assets/icon/OCR_icon/
+    - .env
+    - lib/assets/icon/
+    - lib/assets/svg/
+    - lib/assets/icon/active_navbar/
+    - lib/assets/icon/inactive_navbar/
+    - lib/assets/icon/home_icon/
+    - lib/assets/icon/add_icon/
+    - lib/assets/icon/components_icon/
+    - lib/assets/icon/popup_icon/
+    - lib/assets/icon/setting_icon/
+    - lib/assets/icon/transaction_icon/
+    - lib/assets/icon/figma_svg/
+    - lib/assets/icon/OCR_icon/
 
   fonts:
     - family: Lato
@@ -78,4 +72,3 @@ flutter:
     - family: HappyMonkey
       fonts:
         - asset: lib/assets/fonts/HappyMonkey.ttf
-
